backend/requirements.txt
Django==4.0.1
environs==9.4.0
djangorestframework==3.13.1
django-phonenumber-field[phonenumbers]==6.0.0
Pillow==9.0.0